% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-artifactory.R
\name{artifactory_envvar}
\alias{artifactory_envvar}
\title{Set Artifactory environmental variables.}
\usage{
artifactory_envvar(key, desc, quiet = TRUE)
}
\arguments{
\item{key}{Environment variable name.}

\item{desc}{Short description, e.g., "API key."}

\item{quiet}{Print status message.}
}
\description{
Adapted from \code{\link[=gitlab_pat]{gitlab_pat()}}.
}
\keyword{internal}
