% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-artifactory.R
\name{artifactory_package_metadata}
\alias{artifactory_package_metadata}
\title{Fetch package metadata using the Artifactory API.}
\usage{
artifactory_package_metadata(
  pkgname,
  repo = artifactory_repo(),
  key = artifactory_key(),
  type = "source"
)
}
\arguments{
\item{pkgname}{The remote package name.}

\item{repo}{Artifactory repository URL (assumed \code{"<host>/<repo>"}).}

\item{key}{Artifactory API key.}

\item{type}{Package type (only \code{"source"} allowed for now).}
}
\value{
A list of package metadata.
}
\description{
Fetch package metadata using the Artifactory API.
}
\keyword{internal}
