% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{available_packages_set}
\alias{available_packages_set}
\alias{available_packages_reset}
\alias{available_packages}
\title{Simpler available.packages}
\usage{
available_packages_set(repos, type, db)

available_packages_reset()

available_packages(repos = getOption("repos"), type = getOption("pkgType"))
}
\arguments{
\item{repos}{
    character vector, the base URL(s) of the repositories to use.
  }

\item{type}{
    character string, indicate which type of packages: see
    \code{\link[utils]{install.packages}}.

    If \code{type = "both"} this will use the source repository.
  }
}
\description{
This is mostly equivalent to \code{\link[utils:available.packages]{utils::available.packages()}} however it also
caches the full result. Additionally the cache can be assigned explicitly with
\code{\link[=available_packages_set]{available_packages_set()}} and reset (cleared) with \code{\link[=available_packages_reset]{available_packages_reset()}}.
}
\seealso{
\code{\link[utils:available.packages]{utils::available.packages()}} for full documentation on the output format.
}
\keyword{internal}
