% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{bioc_install_repos}
\alias{bioc_install_repos}
\title{Deduce the URLs of the BioConductor repositories}
\usage{
bioc_install_repos()
}
\value{
A named character vector of the URLs of the
BioConductor repositories, appropriate for the current
R version.
}
\description{
Deduce the URLs of the BioConductor repositories
}
