% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-git.R
\name{git}
\alias{git}
\title{Install a package from a git repository}
\usage{
git(url, subdir = NULL, branch = NULL, args = character(0), ...)
}
\arguments{
\item{url}{Location of package. The url should point to a public or
private repository.}

\item{subdir}{A sub-directory within a git repository that may
contain the package we are interested in installing.}

\item{branch}{Name of branch or tag to use, if not master.}

\item{args}{DEPRECATED. A character vector providing extra arguments to
pass on to git.}

\item{...}{passed on to \code{install.packages}}
}
\description{
It is vectorised so you can install multiple packages with
a single command. You do not need to have git installed.
}
\examples{
\dontrun{
install_git("git://github.com/hadley/stringr.git")
install_git("git://github.com/hadley/stringr.git", branch = "stringr-0.2")
}
}

