% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-git.R
\name{git_anon_url}
\alias{git_anon_url}
\title{Anonymize a git-style url}
\usage{
git_anon_url(url)
}
\arguments{
\item{url}{A \code{character} vector of urls to parse}
}
\description{
Strip a url of user-specific username and password if embedded as part of a
url string.
}
