% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-git.R
\name{git_censored_url}
\alias{git_censored_url}
\title{Censor user password in a git-style url}
\usage{
git_censored_url(url)
}
\arguments{
\item{url}{A \code{character} vector of urls to parse}
}
\description{
If a password is provided as part of a url string, censor the url string,
replacing the password with a series of asterisks.
}
