% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-git.R
\name{git_credentials}
\alias{git_credentials}
\title{Specify git credentials to use}
\usage{
git_credentials()
}
\description{
The global option \code{remotes.git_credentials} is used to set the git
credentials.
}
\keyword{internal}
