% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-github.R
\name{github_pull}
\alias{github_pull}
\alias{github_release}
\title{GitHub references}
\usage{
github_pull(pull)

github_release()
}
\arguments{
\item{pull}{Character string specifying the pull request to install}
}
\description{
Use as \code{ref} parameter to \code{\link[=install_github]{install_github()}}.
Allows installing a specific pull request or the latest release.
}
\examples{
github_pull("42")
}
\seealso{
\code{\link[=install_github]{install_github()}}
}
