% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-gitlab.R
\name{gitlab_pat}
\alias{gitlab_pat}
\title{Retrieve GitLab personal access token.}
\usage{
gitlab_pat(quiet = TRUE)
}
\description{
A GitLab personal access token
Looks in env var \code{GITLAB_PAT}
}
\keyword{internal}
