% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-gitlab.R
\name{gitlab_pull}
\alias{gitlab_pull}
\alias{gitlab_release}
\title{GitLab references}
\usage{
gitlab_pull(pull)

gitlab_release()
}
\arguments{
\item{pull}{The pull request to install}
}
\description{
Use as \code{ref} parameter to \code{\link[=install_gitlab]{install_gitlab()}}.
Allows installing a specific pull request or the latest release.
}
\seealso{
\code{\link[=install_gitlab]{install_gitlab()}}
}
