% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-bioc.R
\name{install_bioc}
\alias{install_bioc}
\title{Install a package from a Bioconductor repository}
\usage{
install_bioc(repo, mirror = getOption("BioC_git",
  "https://git.bioconductor.org/packages"), git = c("auto", "git2r",
  "external"), ..., quiet = FALSE)
}
\arguments{
\item{repo}{Repository address in the format
\code{[username:password@][release/]repo[#commit]}. Valid values for
the release are \sQuote{devel},
\sQuote{release} (the default if none specified), or numeric release
numbers (e.g. \sQuote{3.3}).}

\item{mirror}{The bioconductor git mirror to use}

\item{git}{Whether to use the \code{git2r} package, or an external
git client via system. Default is \code{git2r} if it is installed,
otherwise an external git installation.}

\item{...}{Other arguments passed on to \code{\link{install}}}
}
\description{
This function requires \code{git} to be installed on your system in order to
be used.
}
\details{
It is vectorised so you can install multiple packages with
a single command.

'
}
\examples{
\dontrun{
install_bioc("SummarizedExperiment")
install_bioc("release/SummarizedExperiment")
install_bioc("3.3/SummarizedExperiment")
install_bioc("SummarizedExperiment#abc123")
install_bioc("user:password@release/SummarizedExperiment")
install_bioc("user:password@devel/SummarizedExperiment")
install_bioc("user:password@SummarizedExperiment#abc123")
}
}
\seealso{
Other package installation: \code{\link{install_cran}},
  \code{\link{install_gitlab}},
  \code{\link{install_version}}
}
