% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-gitlab.R
\name{install_gitlab}
\alias{install_gitlab}
\title{Attempts to install a package directly from GitLab.}
\usage{
install_gitlab(repo, ref = "master", subdir = NULL,
  auth_token = gitlab_pat(), host = "www.gitlab.com", ...)
}
\arguments{
\item{repo}{Repository address in the format
\code{username/repo[/subdir][@ref|#merge]}. Alternatively, you can
specify \code{subdir} and/or \code{ref} using the respective parameters
(see below); if both is specified, the values in \code{repo} take
precedence.}

\item{ref}{Desired git reference. Could be a commit, tag, or branch
name, or a call to \code{\link{gitlab_merge}}. Defaults to \code{"master"}.}

\item{subdir}{subdirectory within repo that contains the R package.}

\item{auth_token}{To install from a private repo, generate a personal
access token (PAT) in \url{https://gitlab.com/profile/personal_access_tokens} and
supply to this argument. This is safer than using a password because
you can easily delete a PAT without affecting any others. Defaults to
the \code{GITLAB_PAT} environment variable.}

\item{host}{GitLab API host to use. Override with your GitLab hostname, for 
example, \code{"gitlab.hostname.com/api/v3"}. #check this}

\item{...}{Other arguments passed on to \code{install.packages}.}
}
\description{
This function is vectorised on \code{repo} so you can install multiple
packages in a single command.
}
\details{
Attempting to install from a source repository that uses submodules
raises a warning. Because the zipped sources provided by GitLab do not
include submodules, this may lead to unexpected behaviour or compilation
failure in source packages. In this case, cloning the repository manually
may yield better results.
}
\examples{
\dontrun{
install_gitlab("ConorIA/roxygen")
install_gitlab("wch/ggplot2")
install_gitlab(c("rstudio/httpuv", "rstudio/shiny"))
install_gitlab(c("hadley/httr@v0.4", "klutometis/roxygen#142",
  "mfrasca/r-logging/pkg"))

# To install from a private repo, use auth_token with a token
# from https://gitlab.com/settings/applications. You only need the
# repo scope. Best practice is to save your PAT in env var called
# GITHUB_PAT.
install_gitlab("hadley/private", auth_token = "abc")

}
}
\seealso{
\code{\link{gitlab_merge}}
}

