% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{installgithub}
\alias{installgithub}
\alias{installgithub-package}
\title{Install R Packages from 'GitHub'}
\description{
Download and install R packages stored in GitHub repositories. This package is a lightweight replacement of `devtools::install_github()`.
}

