% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-git.R
\name{parse_git_url}
\alias{parse_git_url}
\title{Extract URL parts from a git-style url}
\usage{
parse_git_url(url)
}
\arguments{
\item{url}{A \code{character} vector of urls to parse}
}
\description{
Although not a full url parser, this expression captures and separates url
protocol (\code{prot}), full authentication prefix (\code{auth}, containing \code{username}
and \code{password}), the host and path (\code{url}) and git reference (\code{ref}).
}
