% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes-package.R
\docType{package}
\name{remotes-package}
\alias{remotes}
\alias{remotes-package}
\title{remotes: R Package Installation from Remote Repositories, Including 'GitHub'}
\description{
Download and install R packages stored in 'GitHub', 'GitLab', 'Bitbucket', 'Bioconductor', or plain 'subversion' or 'git' repositories. This package provides the 'install_*' functions in 'devtools'. Indeed most of the code was copied over from 'devtools'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://remotes.r-lib.org}
  \item \url{https://github.com/r-lib/remotes#readme}
  \item Report bugs at \url{https://github.com/r-lib/remotes/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Jim Hester
  \item Hadley Wickham
  \item Winston Chang
  \item Martin Morgan
  \item Dan Tenenbaum
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Mango Solutions [copyright holder]
}

}
\keyword{internal}
