% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{standardise_dep}
\alias{standardise_dep}
\title{Standardise dependencies using the same logical as \link{install.packages}}
\usage{
standardise_dep(x)
}
\arguments{
\item{x}{The dependencies to standardise.
A character vector (selecting from "Depends", "Imports",
"LinkingTo", "Suggests", or "Enhances"), or a logical vector.

\code{TRUE} is shorthand for "Depends", "Imports", "LinkingTo" and
"Suggests". \code{NA} is shorthand for "Depends", "Imports" and "LinkingTo"
and is the default. \code{FALSE} is shorthand for no dependencies.}
}
\description{
Standardise dependencies using the same logical as \link{install.packages}
}
\seealso{
\url{http://r-pkgs.had.co.nz/description.html#dependencies} for
additional information on what each dependency type means.
}
\keyword{internal}
