% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-version.R
\name{version_satisfies_criteria}
\alias{version_satisfies_criteria}
\title{Check package version with a given criteria}
\usage{
version_satisfies_criteria(to_check, criteria)
}
\arguments{
\item{to_check}{Package version}

\item{criteria}{Criteria used to verify the required package version}
}
\value{
Vector with logical values
}
\description{
Check package version with a given criteria
}
