% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-artifactory.R
\name{artifactory_remote}
\alias{artifactory_remote}
\title{Create a new Artifactory remote}
\usage{
artifactory_remote(
  package,
  repo = artifactory_repo(),
  auth_token = artifactory_key()
)
}
\arguments{
\item{repo}{Artifactory repo URL, e.g., \code{"company.jfrog.io/artifactory/cran-local"}.}
}
\description{
This is an internal function to create a new \code{artifactory_remote}.
}
\keyword{internal}
