% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-artifactory.R
\name{artifactory_repo}
\alias{artifactory_repo}
\alias{artifactory_user}
\alias{artifactory_key}
\title{Set Artifactory environmental variables.}
\usage{
artifactory_repo(quiet = TRUE)

artifactory_user(quiet = TRUE)

artifactory_key(quiet = TRUE)
}
\arguments{
\item{quiet}{XXSD}
}
\description{
DFGHJKL
}
\section{Functions}{
\itemize{
\item \code{artifactory_repo}: AAA

\item \code{artifactory_user}: XXX

\item \code{artifactory_key}: YYY
}}

