% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-gitlab.R
\name{gitlab_merge}
\alias{gitlab_merge}
\alias{gitlab_release}
\title{GitLab references}
\usage{
gitlab_merge(merge)

gitlab_release()
}
\arguments{
\item{merge}{The merge request to install}
}
\description{
Use as \code{ref} parameter to \code{\link{install_gitlab}}.
Allows installing a specific merge request or the latest release.
}
\seealso{
\code{\link{install_gitlab}}
}

