% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-gitlab.R
\name{install_gitlab}
\alias{install_gitlab}
\title{Instlal a package from GitLab}
\usage{
install_gitlab(repo, auth_token = gitlab_pat(), host = "gitlab.com", ...)
}
\arguments{
\item{repo}{Repository address in the format
\code{username/repo[/subdir][@ref]}.}

\item{auth_token}{To install from a private repo, generate a personal
access token (PAT) in \url{https://github.com/settings/applications} and
supply to this argument. This is safer than using a password because
you can easily delete a PAT without affecting any others. Defaults to
the \code{GITHUB_PAT} environment variable.}

\item{host}{GitLab API host to use. Override with your GitLab enterprise
hostname, for example, \code{"gitlab.hostname.com"}.}

\item{...}{Other arguments passed on to \code{\link[utils]{install.packages}}.}
}
\description{
This function is vectorised on \code{repo} so you can install multiple
packages in a single command. Like other remotes the repository will skip
installation if `force == FALSE` (the default) and the remote state has
not changed since the previous installation.
}
\examples{
\dontrun{
install_github("jimhester/covr")
}
}
\seealso{
Other package installation: \code{\link{install_cran}},
  \code{\link{install_version}}
}
