% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{standardise_dep}
\alias{standardise_dep}
\title{Standardise dependencies using the same logical as \link{install.packages}}
\usage{
standardise_dep(x)
}
\arguments{
\item{x}{The dependencies to standardise.
A character vector (selecting from "Depends", "Imports",
"LinkingTo", "Suggests", or "Enhances"), or a logical vector.

\code{TRUE} is shorthand for "Depends", "Imports", "LinkingTo" and
"Suggests". \code{NA} is shorthand for "Depends", "Imports" and "LinkingTo"
and is the default. \code{FALSE} is shorthand for no dependencies.

The value "soft" means the same as \code{TRUE}, "hard" means the same as \code{NA}.

Any additional values that don't match one of the standard dependency
types are filtered out.}
}
\description{
Standardise dependencies using the same logical as \link{install.packages}
}
\seealso{
\url{https://r-pkgs.org/description.html} for
additional information on what each dependency type means.
}
\keyword{internal}
