% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_requirements.R
\name{system_requirements}
\alias{system_requirements}
\title{Query the system requirements for a package (and its dependencies)}
\usage{
system_requirements(package, os, os_release)
}
\arguments{
\item{package}{The package name to query}

\item{os, os_release}{The operating system and operating system release version, see
\url{https://github.com/rstudio/r-system-requirements#operating-systems} for the
list of supported operating systems.}
}
\description{
Query the system requirements for a package (and its dependencies)
}
